Create unified test file for galvanize, pixilate, and refrigerate
